Replace unsigned int with NSUInteger. Makes building with Xcode 6 & iOS 8 SDK happy. Most of the time, it was working with NSArray length & indexes, or NSData length, so using NSUInteger makes sense.

+++ Frameworks/SnoizeMIDI/SMSysExSendRequest.h
@@ -35,9 +35,9 @@
     // Returns YES if the request was cancelled before it was finished sending; NO if it was already finished.
     // In either case, SMSysExSendRequestFinishedNotification will be posted.
 
-- (unsigned int)bytesRemaining;
-- (unsigned int)totalBytes;
-- (unsigned int)bytesSent;
+- (NSUInteger)bytesRemaining;
+- (NSUInteger)totalBytes;
+- (NSUInteger)bytesSent;
 - (BOOL)wereAllBytesSent;
 
 @end
